+++ etc/blocks/monetization-test/model/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @milaboratories/milaboratories.monetization-test.model
 
+## 1.2.16
+
+### Patch Changes
+
+- @platforma-sdk/model@1.77.4
+
 ## 1.2.15
 
 ### Patch Changes