+++ etc/blocks/monetization-test/ui/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @milaboratories/milaboratories.monetization-test.ui
 
+## 1.3.22
+
+### Patch Changes
+
+- @platforma-sdk/model@1.77.4
+- @milaboratories/milaboratories.monetization-test.model@1.2.16
+- @platforma-sdk/ui-vue@1.77.4
+
 ## 1.3.21
 
 ### Patch Changes