+++ etc/blocks/monetization-test/workflow/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @milaboratories/milaboratories.monetization-test.workflow
 
+## 1.2.20
+
+### Patch Changes
+
+- Updated dependencies [030e8c2]
+  - @platforma-sdk/workflow-tengo@5.25.0
+
 ## 1.2.19
 
 ### Patch Changes