+++ etc/blocks/ui-examples/model/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @milaboratories/milaboratories.ui-examples.model
 
+## 1.6.16
+
+### Patch Changes
+
+- @platforma-sdk/model@1.77.4
+
 ## 1.6.15
 
 ### Patch Changes