+++ etc/blocks/ui-examples/ui/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @milaboratories/milaboratories.ui-examples.ui
 
+## 1.11.16
+
+### Patch Changes
+
+- @platforma-sdk/model@1.77.4
+- @milaboratories/milaboratories.ui-examples.model@1.6.16
+
 ## 1.11.15
 
 ### Patch Changes